requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
cssselect==1.2.0
Cython==0.29.32
fake-useragent==0.1.13
idna==3.4
importlib-metadata==5.0.0
lxml==4.9.1
parse==1.19.0
psycopg2-binary==2.9.5
pyaes==1.6.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
python-dotenv==0.20.0
requests==2.28.1
selectolax==0.3.8
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
w3lib==2.0.1
websockets==10.4
zipp==3.10.0
